Use ready event for E2E hydration
