--- app/assets/entry.e2e.ts
@@ -1,7 +1,0 @@
-import { startApp } from "./start.ts";
-
-let app = startApp();
-
-await app.ready();
-
-document.documentElement.dataset.remixReady = "true";
